pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "synthia-email-node"
version = "0.1.0"
description = "Phase 1 Synthia email node runtime"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115,<1",
  "httpx>=0.28,<0.29",
  "paho-mqtt>=2.1,<3",
  "pydantic-settings>=2.6,<3",
  "uvicorn>=0.34,<0.35",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9",
  "pytest-asyncio>=0.25,<0.26",
]

[tool.pytest.ini_options]
addopts = "-q"
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.setuptools]
py-modules = [
  "config",
  "core_client",
  "logging_utils",
  "main",
  "models",
  "mqtt",
  "service",
  "state_store",
  "version",
]

[tool.setuptools.packages.find]
where = ["src"]
